# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crossword Helper API
  version: 2.0.0
  description: |
    REST API for crossword construction assistance tools including pattern matching,
    grid autofill, theme word placement, and wordlist management.

    **Architecture**: Flask backend delegating to CLI via subprocess for single source of truth.

    **Features**:
    - Pattern search across wordlists
    - Grid numbering and validation
    - Entry normalization (crossword conventions)
    - CSP-based autofill with progress tracking
    - Theme word placement suggestions
    - Strategic black square placement
    - Pause/resume autofill with user edits
    - Wordlist management and import

  contact:
    name: Crossword Helper Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server (this is a local-only tool; no other deployment exists)
...
paths:
...
  /wordlists:
    get:
      tags:
        - Wordlist Management
      summary: List wordlists
      description: List all available wordlists with metadata and categories
      operationId: listWordlists
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category (core, themed, custom, etc.)
      responses:
        '200':
          description: Wordlist metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordlistListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /wordlists/{name}:
    get:
      tags:
        - Wordlist Management
      summary: Get wordlist
      description: Get specific wordlist content and metadata
      operationId: getWordlist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Wordlist name (can include category path, e.g., "core/common_3_letter")
        - name: stats
          in: query
          schema:
            type: boolean
            default: false
          description: Include word statistics
      responses:
        '200':
          description: Wordlist content and metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordlistResponse'
        '404':
          description: Wordlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Wordlist Management
      summary: Create wordlist
      description: Create a new wordlist with words and metadata
      operationId: createWordlist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - words
              properties:
                words:
                  type: array
                  items:
                    type: string
                metadata:
                  $ref: '#/components/schemas/WordlistMetadata'
      responses:
        '201':
          description: Wordlist created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  word_count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Wordlist Management
      summary: Update wordlist
      description: Update existing wordlist (replace, add, or remove words)
      operationId: updateWordlist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                words:
                  type: array
                  items:
                    type: string
                  description: Replace all words
                add_words:
                  type: array
                  items:
                    type: string
                  description: Add specific words
                remove_words:
                  type: array
                  items:
                    type: string
                  description: Remove specific words
                metadata:
                  $ref: '#/components/schemas/WordlistMetadata'
      responses:
        '200':
          description: Wordlist updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Wordlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Wordlist Management
      summary: Delete wordlist
      description: Delete a wordlist file
      operationId: deleteWordlist
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wordlist deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Wordlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /wordlists/{name}/stats:
    get:
      tags:
        - Wordlist Management
      summary: Get wordlist statistics
      description: Get detailed statistics for a wordlist
      operationId: getWordlistStats
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wordlist statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordlistStats'
        '404':
          description: Wordlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /wordlists/search:
    post:
      tags:
        - Wordlist Management
      summary: Search for words across wordlists
      description: |
        Pattern search scoped to specific wordlists (or all, if omitted),
        with results grouped by word and annotated with every source
        wordlist that contains it.
      operationId: searchWordlists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pattern
              properties:
                pattern:
                  type: string
                  example: "?A?E"
                wordlists:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Search results grouped by word
          content:
            application/json:
              schema:
                type: object
                properties:
                  pattern:
                    type: string
                  total_matches:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        word:
                          type: string
                        sources:
                          type: array
                          items:
                            type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /wordlists/import:
    post:
      tags:
        - Wordlist Management
      summary: Import wordlist
      description: Import wordlist from text content or URL
      operationId: importWordlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                content:
                  type: string
                  description: Newline-separated words
                url:
                  type: string
                  format: uri
                  description: URL to fetch wordlist from
                category:
                  type: string
                  default: imports
                metadata:
                  $ref: '#/components/schemas/WordlistMetadata'
      responses:
        '201':
          description: Wordlist imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
                  word_count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: URL import not implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================================
  # Constraint Analysis
  # ============================================================================
...
components:
  schemas:
...
    WordlistListResponse:
      type: object
      properties:
        wordlists:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              category:
                type: string
              word_count:
                type: integer
              metadata:
                $ref: '#/components/schemas/WordlistMetadata'
        categories:
          type: object
          additionalProperties:
            type: integer
          description: Map of category → count
        tags:
          type: object
          additionalProperties:
            type: integer
          description: Map of tag → count
    WordlistResponse:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/WordlistMetadata'
        words:
          type: array
          items:
            type: string
        stats:
          $ref: '#/components/schemas/WordlistStats'
    WordlistMetadata:
      type: object
      properties:
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        source:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    WordlistStats:
      type: object
      properties:
        total_words:
          type: integer
        length_distribution:
          type: object
          additionalProperties:
            type: integer
          description: Map of length → count
        avg_length:
          type: number
          format: float
        min_length:
          type: integer
        max_length:
          type: integer
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    Error:
      description: |
        Two shapes are actually in use (backend/api/errors.py vs. ad-hoc
        jsonify calls) — there is no single canonical error schema.
      oneOf:
        - $ref: '#/components/schemas/ErrorEnvelope'
        - $ref: '#/components/schemas/ErrorFlat'
    ErrorEnvelope:
      type: object
      description: |
        Shape returned by the shared handle_error() helper — most endpoints
        (core operations, grid/theme mutation routes, pause/resume).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorFlat:
      type: object
      description: |
        Shape returned by endpoints that build jsonify() directly instead of
        going through handle_error() — constraint analysis, wordlist search,
        and several validation branches in grid/theme routes.
      properties:
        error:
          type: string
          description: Error message
        validation:
          $ref: '#/components/schemas/ValidationResult'
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security: []  # No authentication scheme exists — this is a local-only tool